move attr class mixin into unified reopenClass
